requirements.txt
Django==1.6
Django-Select2==4.2.2
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==3.0.0
South==1.0
Unidecode==0.04.18
argparse==1.2.1
distribute==0.6.24
django-appconf==1.0.1
django-axes==1.3.6
django-compressor==1.4
django-debug-toolbar==1.2.1
django-nvd3==0.7.5
django-polymorphic==0.5.6
django-reversion==1.8.4
django-secure==1.0
django-simple-captcha==0.4.2
django-tables2==0.15.0
django-watson==1.1.3
psycopg2==2.6.1
python-nvd3==0.13.2
python-slugify==0.0.9
six==1.10.0
slugify==0.0.1
sqlparse==0.1.18
wsgiref==0.1.2
